Finally doing the cleanup I shouldve done a while ago, seperating Groups into Controllers and Models
