+++ router/Models/helper.go
@@ -1,0 +1,25 @@
+package models
+
+import (
+	"database/sql"
+	"example.com/m/v2/middlewares"
+)
+
+func Helper_ExecError(r sql.Result, initerr error, noRowsFound_errMsg string) (int64, error) {
+	// CHecks if the initial call to Query has an error
+	if initerr != nil {
+		return 0, initerr
+	}
+
+	// Checks how many rows are affected, and returns nil if there's an error
+	rowsAffected, err := r.RowsAffected()
+	if err != nil {
+		return 0, err
+	}
+
+	if rowsAffected == 0 {
+		return 0, errors.New(noRowsFound_errMsg)
+	}
+
+	return rowsAffected, nil
+}
